requirements.txt
numpy
colormath
python-rtmidi==0.5b1
